requirements.txt
django
psycopg2-binary
gunicorn
supabase
python-dotenv
pyjwt
bcrypt
python-dotenv

# Database Password: Bussin@Industries